Makefile.am: remove all convenience libraries
With a single non-recursive Makefile.am the libtool-convenience libraries
can be avoided and all source files mentioned in the final _SOURCE of
the library.

In src_libsieve_la_SOURCES the subdirectories are ordered alphabetically,
and the files in each subdirectory are ordered alphabetically, too.

Add the *h files to _SOURCES, near the *c files, so that they are included
in the tarball with "make dist".

This leads to a reduction of the generated Makefile.in from Makefile.am
from 44571 to 40350 bytes, and smaller size for the static libsieve.a .

src/sv_emtpy.c can be removed.